+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Support output block in tag line.
 - Support HTML comment line.
+- Support HTML conditional comment line.
 
 ## 0.2.0 - 2022-01-25